% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_plot.R
\name{beta_plot}
\alias{beta_plot}
\title{Draw values from a beta distribution and plot the probability density
function}
\usage{
beta_plot(n = 10000, a = 1, b = 3)
}
\arguments{
\item{n}{number of observations to draw}

\item{a}{non-negative alpha parameter of the beta distribution}

\item{b}{non-negative beta parameter of the beta distribution}
}
\description{
Draw values from a beta distribution and plot the probability density
function
}
